refactor: 옵션 제출 훅 구현
